refactor: decompose MessageList lifecycle and rendering (#16)
